higher ignition resolution by one decimal point

+++ RUthless_ECU/src/include/igncalc.h
@@ -13,8 +13,8 @@
 // Calculate ignition timing
 uint16_t igncalc_get_ign_degree(void);
 
-uint16_t igncalc_ign_time_teeth(uint8_t ign_degree);
-uint32_t igncalc_ign_time_interval(uint8_t ign_degree);
+uint16_t igncalc_ign_time_teeth(uint16_t ign_degree);
+uint32_t igncalc_ign_time_interval(uint16_t ign_degree);
 
 uint32_t igncalc_dwell_degree(void);
 uint8_t igncalc_dwell_correction(void);
